# apps/api/src/questions/dto/questions.dto.ts
export class QuestionCreateDto {
  description: string;
  title: string;
  question: string;
  option1: string;
  option2: string;
  option3: string;
  option4: string;
  answer: string;
}

export class QuestionUpdateDto {
  description: string;
  title: string;
  question: string;
  option1: string;
  option2: string;
  option3: string;
  option4: string;
  answer: string;
}